# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Terminal Server",
    "version": "1.0"
  },
  "paths": {
    "/screenshot": {
      "post": {
        "tags": [
          "Computer Vision"
        ],
        "summary": "Take Screenshot",
        "description": "Captures a screenshot, compresses it, and returns a Base64 string.",
        "operationId": "take_screenshot_screenshot_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/read-screen": {
      "post": {
        "tags": [
          "Computer Vision"
        ],
        "summary": "Read Screen",
        "description": "Extracts text from the screen using OCR.",
        "operationId": "read_screen_read_screen_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/set-current-directory": {
      "post": {
        "tags": [
          "System Control"
        ],
        "summary": "Change Current Directory",
        "description": "Changes the current working directory.",
        "operationId": "change_current_directory_set_current_directory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CDRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/get-current-directory": {
      "post": {
        "tags": [
          "System Control"
        ],
        "summary": "Return Current Directory",
        "description": "Returns the current working directory.",
        "operationId": "return_current_directory_get_current_directory_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/run-command": {
      "post": {
        "tags": [
          "System Control"
        ],
        "summary": "Run Terminal Command",
        "description": "Runs a terminal command and returns its output.",
        "operationId": "run_terminal_command_run_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/start-process": {
      "post": {
        "tags": [
          "System Control"
        ],
        "summary": "Start Process",
        "description": "Starts a long-running process in the background.\nReturns immediately with a success message and a unique process ID.",
        "operationId": "start_process_start_process_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stop-process/{process_id}": {
      "post": {
        "tags": [
          "System Control"
        ],
        "summary": "Stop Process",
        "description": "Stops a running process given its process ID and deletes its log file.",
        "operationId": "stop_process_stop_process__process_id__post",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Process Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/check-process-status/{process_id}": {
      "post": {
        "tags": [
          "System Control"
        ],
        "summary": "Check Process Status",
        "description": "Checks the status of a running process by reading its log file.",
        "operationId": "check_process_status_check_process_status__process_id__post",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Process Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/list-running-processes": {
      "post": {
        "tags": [
          "System Control"
        ],
        "summary": "List Processes",
        "description": "Lists all processes tracked in the running_processes dictionary.",
        "operationId": "list_processes_list_running_processes_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/replace-function": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Replace Function",
        "description": "Replaces the definition of a function in the specified file.",
        "operationId": "replace_function_replace_function_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceFunctionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/replace-text": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Replace Text Endpoint",
        "description": "Replaces a specific text snippet with new text in the file.",
        "operationId": "replace_text_endpoint_replace_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/read-file": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Get File",
        "description": "Reads the content of the specified file.",
        "operationId": "get_file_read_file_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/write-file": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Write File Endpoint",
        "description": "Writes new content to the specified file.",
        "operationId": "write_file_endpoint_write_file_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/append-file": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Append File Endpoint",
        "description": "Appends content to the specified file.",
        "operationId": "append_file_endpoint_append_file_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/read-lines": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Read File Lines Endpoint",
        "description": "Reads a specified number of lines from the file, starting from a given line index.",
        "operationId": "read_file_lines_endpoint_read_lines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadLinesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/read-logs": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Read Shell Logs",
        "description": "Reads the contents of the system log file.",
        "operationId": "read_shell_logs_read_logs_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/read-function": {
      "post": {
        "tags": [
          "Read/Write Files"
        ],
        "summary": "Read Function",
        "description": "Reads the source code of a specific function from a Python file.",
        "operationId": "read_function_read_function_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadFuncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/info": {
      "post": {
        "tags": [
          "System Information"
        ],
        "summary": "Get Host Info",
        "description": "Returns detailed system information about the host machine.",
        "operationId": "get_host_info_info_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/host-resources": {
      "post": {
        "tags": [
          "System Information"
        ],
        "summary": "Get Host Resources",
        "description": "Returns system resource usage (CPU, RAM, Disk).",
        "operationId": "get_host_resources_host_resources_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/docs": {
      "post": {
        "tags": [
          "Api Documentation"
        ],
        "summary": "Get Openapi Spec",
        "description": "Returns the OpenAPI documentation.",
        "operationId": "get_openapi_spec_docs_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/metadata": {
      "post": {
        "tags": [
          "Api Documentation"
        ],
        "summary": "Get Metadata",
        "description": "Provides metadata about the API.",
        "operationId": "get_metadata_metadata_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "post": {
        "tags": [
          "Api Documentation"
        ],
        "summary": "Health Check",
        "description": "Returns the API status and uptime.",
        "operationId": "health_check_health_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "post": {
        "summary": "Root",
        "operationId": "root__post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/docs-json": {
      "get": {
        "summary": "Read Json File",
        "operationId": "read_json_file_docs_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/restart-server": {
      "post": {
        "summary": "Restart Server",
        "description": "Endpoint to restart the API server.",
        "operationId": "restart_server_restart_server_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/privacy-policy": {
      "get": {
        "summary": "Privacy Policy",
        "operationId": "privacy_policy_privacy_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AppendFileRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "filepath",
          "content"
        ],
        "title": "AppendFileRequest"
      },
      "CDRequest": {
        "properties": {
          "directory": {
            "type": "string",
            "title": "Directory"
          }
        },
        "type": "object",
        "required": [
          "directory"
        ],
        "title": "CDRequest"
      },
      "CommandRequest": {
        "properties": {
          "command": {
            "type": "string",
            "title": "Command"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "CommandRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReadFileRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          }
        },
        "type": "object",
        "required": [
          "filepath"
        ],
        "title": "ReadFileRequest"
      },
      "ReadFuncRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "function_name": {
            "type": "string",
            "title": "Function Name"
          }
        },
        "type": "object",
        "required": [
          "filepath",
          "function_name"
        ],
        "title": "ReadFuncRequest"
      },
      "ReadLinesRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "start_line": {
            "type": "integer",
            "title": "Start Line"
          },
          "num_lines": {
            "type": "integer",
            "title": "Num Lines"
          }
        },
        "type": "object",
        "required": [
          "filepath",
          "start_line",
          "num_lines"
        ],
        "title": "ReadLinesRequest"
      },
      "ReplaceFunctionRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "function_name": {
            "type": "string",
            "title": "Function Name"
          },
          "new_function_code": {
            "type": "string",
            "title": "New Function Code"
          }
        },
        "type": "object",
        "required": [
          "filepath",
          "function_name",
          "new_function_code"
        ],
        "title": "ReplaceFunctionRequest"
      },
      "ReplaceTextRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "original_text": {
            "type": "string",
            "title": "Original Text"
          },
          "replacement_text": {
            "type": "string",
            "title": "Replacement Text"
          }
        },
        "type": "object",
        "required": [
          "filepath",
          "original_text",
          "replacement_text"
        ],
        "title": "ReplaceTextRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WriteFileRequest": {
        "properties": {
          "filepath": {
            "type": "string",
            "title": "Filepath"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "filepath",
          "content"
        ],
        "title": "WriteFileRequest"
      }
    }
  },
  "servers": [
    {
      "url": "https://api.armand0e.online",
      "description": "Production server"
    }
  ]
}